Adds general environment from which every environment inherits
