+++ pointRobot/envs/pointRobotEnv.py
@@ -2,15 +2,9 @@ import numpy as np
 import time
 from abc import abstractmethod
 
-from scipy.integrate import odeint
+from planarCommon.planarEnv import PlanarEnv
 
-from gym import core
-from gym.utils import seeding
-
-
-class PointRobotEnv(core.Env):
-
-    metadata = {"render.modes": ["human", "rgb_array"], "video.frames_per_second": 15}
+class PointRobotEnv(PlanarEnv):
 
     MAX_VEL = 10
     MAX_POS = 10
@@ -33,13 +27,6 @@ class PointRobotEnv(core.Env):
     @abstractmethod
     def setSpaces(self):
         pass
-
-    def dt(self):
-        return self._dt
-
-    def seed(self, seed=None):
-        self.np_random, seed = seeding.np_random(seed)
-        return [seed]
 
     def reset(self, pos=None, vel=None):
         if not isinstance(pos, np.ndarray) or not pos.size == self._n:
@@ -72,12 +59,6 @@ class PointRobotEnv(core.Env):
     def continuous_dynamics(self, x, t):
         pass
 
-    def integrate(self):
-        x0 = self.state[0 : 2 * self._n]
-        t = np.arange(0, 2 * self._dt, self._dt)
-        ynext = odeint(self.continuous_dynamics, x0, t)
-        return ynext[1]
-
     def render(self, mode="human"):
         from gym.envs.classic_control import rendering
 
@@ -103,8 +84,3 @@ class PointRobotEnv(core.Env):
         time.sleep(self.dt())
 
         return self.viewer.render(return_rgb_array=mode == "rgb_array")
-
-    def close(self):
-        if self.viewer:
-            self.viewer.close()
-            self.viewer = None
